requirements.txt
chunkr-ai
openai
numpy
faiss-cpu
python-dotenv
langchain
langchain-openai
langchain-community
gradio
langgraph
